+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.2] - 2025-06-??
+
+- Add built-in features to enable `OpenTelemetry` logging for all web requests
+  and exceptions, based on the examples provided in the
+  [BlackSheep-Examples](https://github.com/Neoteroi/BlackSheep-Examples/tree/main/otel)
+  repository.
+
 ## [2.3.1] - 2025-05-20 :racehorse:
 
 - Add support for [`PyPy`](https://pypy.org/), adding a pure-Python fallback